+++ frontend/src/components/fields/DateTimeSettingsEdit.vue
@@ -5,7 +5,7 @@ import { useI18n } from "vue-i18n";
 
 import DateTimeInput from "@/components/shared/DateTimeInput.vue";
 import type { DateTimeSettings, FieldType, UpdateFieldTypePayload } from "@/types/fields";
-import { updateFieldType } from "@/api/fields";
+import { updateFieldType } from "@/api/fields/types";
 import { useForm } from "@/forms";
 
 const { t } = useI18n();
