+++ frontend/src/components/fields/DeleteFieldType.vue
@@ -4,7 +4,7 @@ import { computed, ref } from "vue";
 import { useI18n } from "vue-i18n";
 
 import type { FieldType } from "@/types/fields";
-import { deleteFieldType } from "@/api/fields";
+import { deleteFieldType } from "@/api/fields/types";
 
 const { t } = useI18n();
